feat: past sessions display in a row at the bottom of the live page
